requirements.txt
amqp==2.5.0
anyjson==0.3.3
backports.csv==1.0.7
billiard==3.6.0.0
celery==4.4.0rc2
celery-with-redis==3.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.4
django-celery==3.3.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-guardian==2.0.0
django-import-export==1.2.0
django-redis==4.10.0
django-stdimage==5.0.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
dnspython==1.16.0
drf-writable-nested==0.5.1
et-xmlfile==1.0.1
eventlet==0.25.0
future==0.17.1
geoip2==2.9.0
greenlet==0.4.15
httplib2==0.13.1
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
kombu==4.6.3
MarkupSafe==1.1.1
maxminddb==1.4.1
monotonic==1.5
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.1.0
progressbar2==3.42.0
PyJWT==1.7.1
python-utils==2.3.0
pytz==2019.2
PyYAML==5.1.2
redis==3.3.6
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
